+++ src/test/java/inflearn/springadvance/proxy/proxyfactory/ProxyFactoryTest.java
@@ -1,12 +1,16 @@
 package inflearn.springadvance.proxy.proxyfactory;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import inflearn.springadvance.proxy.common.advice.TimeAdvice;
+import inflearn.springadvance.proxy.common.service.ConcreteService;
 import inflearn.springadvance.proxy.common.service.ServiceImpl;
 import inflearn.springadvance.proxy.common.service.ServiceInterface;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.aop.framework.ProxyFactory;
+import org.springframework.aop.support.AopUtils;
 
 @Slf4j
 public class ProxyFactoryTest {
@@ -22,8 +26,32 @@ public class ProxyFactoryTest {
         proxyFactory.addAdvice(timeAdvice);
 
         ServiceInterface proxy = (ServiceInterface) proxyFactory.getProxy();
+        proxy.save();
         log.info("targetClass={}", target.getClass());
         log.info("proxyClass={}", proxy.getClass());
 
+        assertThat(AopUtils.isAopProxy(proxy)).isTrue();
+        assertThat(AopUtils.isJdkDynamicProxy(proxy)).isTrue();
+        assertThat(AopUtils.isCglibProxy(proxy)).isFalse();
+
+    }
+
+    @Test
+    @DisplayName("구체 클래스가 있으면 CGLIB 동적 프록시 사용")
+    void concrete_proxy_cglib(){
+
+        ConcreteService target = new ConcreteService();
+        ProxyFactory proxyFactory = new ProxyFactory(target);
+
+        proxyFactory.addAdvice(new TimeAdvice());
+        ConcreteService proxy = (ConcreteService) proxyFactory.getProxy();
+
+        proxy.call();
+        log.info("targetClass={}", target.getClass());
+        log.info("proxyClass={}", proxy.getClass());
+
+        assertThat(AopUtils.isAopProxy(proxy)).isTrue();
+        assertThat(AopUtils.isJdkDynamicProxy(proxy)).isFalse();
+        assertThat(AopUtils.isCglibProxy(proxy)).isTrue();
     }
 }